Extracts images from a video file with ffmpeg

Uses ffmpeg systems calls to extract images from a video.

vid.to.images(vid.path = NULL, out.dir = NULL, overwrite = FALSE, qual = 50)

Arguments

vid.path
Character; path of video file to be processed.

out.dir
character; directory path in which to store images.

overwrite
logical; should path described by 'out.dir' be overwritten if it exhists.

qual
numeric; the quality of the jpeg images to be rendered from 1-100%. Defaults to 50%.

Value

Extracts all the images of the video and saves them to an "images" directory with appended number sequence

See also

Examples

vid.to.images,Extracts images from a video file with ffmpeg,Extracts images from a video file with ffmpeg — vid.to.images,vid.to.images.Rd,character(0),vid.to.images,character(0),list(title = "Description", contents = "
Uses ffmpeg systems calls to extract images from a video.
"),list(description = "Uses ffmpeg systems calls to extract images from a video."),vid.to.images(vid.path = NULL, out.dir = NULL, overwrite = FALSE, qual = 50),list(list(name = "vid.path", description = "
Character; path of video file to be processed.
"), list(name = "out.dir", description = "
character; directory path in which to store images.
"), list(name = "overwrite", description = "
logical; should path described by 'out.dir' be overwritten if it exhists.
"), list(name = "qual", description = "
numeric; the quality of the jpeg images to be rendered from 1-100%. Defaults to 50%.
")),TRUE,

#make a video with animation package
# \donttest{
require(animation)

#> Loading required package: animation

fun <- function(){
y <- sin(1:50)
x <- 1:50
for(i in 1:50) {
  plot(x[i],y[i],col="red",xlim=c(0,50),ylim=range(y))
  animation::ani.pause()
  }
}
animation::saveVideo(fun(),video.name=paste0(tempdir(),"/wave.mp4"),interval = 0.2)

#> Executing: 'ffmpeg' -y -framerate 5 -i Rplot%d.png -pix_fmt yuv420p wave.mp4

#> 
#> 
#> Video has been created at: /private/var/folders/z3/_825pg0s3jvf0hb_q8kzmg5h0000gn/T/Rtmps2Jdnx/wave.mp4

#create directory in which to store images
dir.create(paste0(tempdir(),"/images"))
vid.to.images(vid.path=paste0(tempdir(),"/wave.mp4"),
out.dir= paste0(tempdir(),"/images"),qual=100)

#see the images in the "images" subdirectory
list.files( paste0(tempdir(),"/images"))

#>  [1] "wave.mp4_00001.jpg" "wave.mp4_00002.jpg" "wave.mp4_00003.jpg"
#>  [4] "wave.mp4_00004.jpg" "wave.mp4_00005.jpg" "wave.mp4_00006.jpg"
#>  [7] "wave.mp4_00007.jpg" "wave.mp4_00008.jpg" "wave.mp4_00009.jpg"
#> [10] "wave.mp4_00010.jpg" "wave.mp4_00011.jpg" "wave.mp4_00012.jpg"
#> [13] "wave.mp4_00013.jpg" "wave.mp4_00014.jpg" "wave.mp4_00015.jpg"
#> [16] "wave.mp4_00016.jpg" "wave.mp4_00017.jpg" "wave.mp4_00018.jpg"
#> [19] "wave.mp4_00019.jpg" "wave.mp4_00020.jpg" "wave.mp4_00021.jpg"
#> [22] "wave.mp4_00022.jpg" "wave.mp4_00023.jpg" "wave.mp4_00024.jpg"
#> [25] "wave.mp4_00025.jpg" "wave.mp4_00026.jpg" "wave.mp4_00027.jpg"
#> [28] "wave.mp4_00028.jpg" "wave.mp4_00029.jpg" "wave.mp4_00030.jpg"
#> [31] "wave.mp4_00031.jpg" "wave.mp4_00032.jpg" "wave.mp4_00033.jpg"
#> [34] "wave.mp4_00034.jpg" "wave.mp4_00035.jpg" "wave.mp4_00036.jpg"
#> [37] "wave.mp4_00037.jpg" "wave.mp4_00038.jpg" "wave.mp4_00039.jpg"
#> [40] "wave.mp4_00040.jpg" "wave.mp4_00041.jpg" "wave.mp4_00042.jpg"
#> [43] "wave.mp4_00043.jpg" "wave.mp4_00044.jpg" "wave.mp4_00045.jpg"
#> [46] "wave.mp4_00046.jpg" "wave.mp4_00047.jpg" "wave.mp4_00048.jpg"
#> [49] "wave.mp4_00049.jpg" "wave.mp4_00050.jpg"

#clean up
unlink(paste0(tempdir(),"/images"),recursive=TRUE)
# }

,list(list(title = "Value", contents = "
Extracts all the images of the video and saves them to an \"images\" directory with appended number sequence
", slug = "value"), list(title = "See also", contents = "

", slug = "see-also")),2021,list(name = "trackter", version = "0.1.1", authors = "Christopher Kenaley"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "trackter"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  \n    Articles\n     \n    \n  \n  
\n    
\n      tutorial\n    
\n  
\n
\n
\n  Changelog\n
", 
    right = ""),list(bootswatch = "cerulean", .present = TRUE),list(version = "1.6.1"),FALSE