Extracts images from a video file with ffmpeg

Extract images from video file using ffmpegs flexible video filters and codecs

vid.to.images2(
  vid.path = NULL,
  out.dir = NULL,
  overwrite = FALSE,
  filt = NULL,
  codec = NULL,
  silent = TRUE
)

Arguments

vid.path
character; path of video file to be processed.

out.dir
character; directory path in which to store images.

overwrite
logical; should path described by 'out.dir' be overwritten if it exists.

filt
character; video filter that should be applied to ffmpeg operation. See https://ffmpeg.org/ffmpeg-filters.html

codec
character; video codec to apply in ffmpeg operation

silent
logical; should output of system call for ffmpeg operation be suppressed.

Value

Extracts all the images of the video and saves them to an "images" directory with appended number sequence

Details

Particularly useful for resizing images

See also

Examples

#make a video with animation package
# \donttest{
fun <- function(){
y <- sin(1:50)
x <- 1:50
for(i in 1:50) {
  plot(x[i],y[i],col="red",xlim=c(0,50),ylim=range(y))
  animation::ani.pause()
  }
}
animation::saveVideo(fun(),video.name=paste0(tempdir(),"/wave.mp4"),interval = 0.2)

#> Executing: 'ffmpeg' -y -framerate 5 -i Rplot%d.png -pix_fmt yuv420p wave.mp4

#> 
#> 
#> Video has been created at: /private/var/folders/5c/s3c2hjfd7blbqwwpl1q10wxm0000gy/T/RtmpH0VL1s/wave.mp4

#reduce the image images to 200 px wide maintaining aspect ratio
#notice the spaces at the beginning/end of string
filt.red <- " -vf scale=200:-1 "
c <- " -c:v libx264 "
dir.create(paste0(tempdir(),"/images"))
vid.to.images2(vid.path=paste0(tempdir(),"/wave.mp4"),
out.dir=paste0(tempdir(),"/images"),filt=filt.red,codec=NULL)

#see the images in the "images" directory
list.files( paste0(tempdir(),"/images"))

#>  [1] "wave.mp4_00001.jpg" "wave.mp4_00002.jpg" "wave.mp4_00003.jpg"
#>  [4] "wave.mp4_00004.jpg" "wave.mp4_00005.jpg" "wave.mp4_00006.jpg"
#>  [7] "wave.mp4_00007.jpg" "wave.mp4_00008.jpg" "wave.mp4_00009.jpg"
#> [10] "wave.mp4_00010.jpg" "wave.mp4_00011.jpg" "wave.mp4_00012.jpg"
#> [13] "wave.mp4_00013.jpg" "wave.mp4_00014.jpg" "wave.mp4_00015.jpg"
#> [16] "wave.mp4_00016.jpg" "wave.mp4_00017.jpg" "wave.mp4_00018.jpg"
#> [19] "wave.mp4_00019.jpg" "wave.mp4_00020.jpg" "wave.mp4_00021.jpg"
#> [22] "wave.mp4_00022.jpg" "wave.mp4_00023.jpg" "wave.mp4_00024.jpg"
#> [25] "wave.mp4_00025.jpg" "wave.mp4_00026.jpg" "wave.mp4_00027.jpg"
#> [28] "wave.mp4_00028.jpg" "wave.mp4_00029.jpg" "wave.mp4_00030.jpg"
#> [31] "wave.mp4_00031.jpg" "wave.mp4_00032.jpg" "wave.mp4_00033.jpg"
#> [34] "wave.mp4_00034.jpg" "wave.mp4_00035.jpg" "wave.mp4_00036.jpg"
#> [37] "wave.mp4_00037.jpg" "wave.mp4_00038.jpg" "wave.mp4_00039.jpg"
#> [40] "wave.mp4_00040.jpg" "wave.mp4_00041.jpg" "wave.mp4_00042.jpg"
#> [43] "wave.mp4_00043.jpg" "wave.mp4_00044.jpg" "wave.mp4_00045.jpg"
#> [46] "wave.mp4_00046.jpg" "wave.mp4_00047.jpg" "wave.mp4_00048.jpg"
#> [49] "wave.mp4_00049.jpg" "wave.mp4_00050.jpg"

#clean up
unlink(paste0(tempdir(),"/images"),recursive=TRUE)
# }